헤더바 컴포넌트화
